Fix selected element has to be an ID and get from page blocks
If we set the element directly, we can not modify it, as its a diferent pointer, so changing the element does change the page but not the elemnet for react managed components.

Using an id, on ecah change we have to look for the element, but its the real updated data always.